commit initial

+++ src/main/java/fr/alteca/poc/service/impl/BlogServiceImpl.java
@@ -1,6 +1,7 @@
 package fr.alteca.poc.service.impl;
 
 import fr.alteca.poc.entity.Blog;
+import fr.alteca.poc.exception.CustomException;
 import fr.alteca.poc.service.BlogService;
 
 import java.util.ArrayList;
@@ -13,25 +14,35 @@ public class BlogServiceImpl implements BlogService {
         return result;
     }
 
-    public List<Blog> getByCriteria(String title, String content, String hashTag) {
+    public List<Blog> getByCriteria(String title, String content, String hashTag) throws CustomException {
+        if (title == null || content == null || hashTag == null)
+            throw new CustomException("Parametre null");
         List<Blog> result = new ArrayList<Blog>();
 
         return result;
     }
 
-    public Blog getById(int id) {
+    public Blog getById(Integer id) throws CustomException {
+        if (id == null)
+            throw new CustomException("Parametre null");
+
         Blog result = new Blog();
 
         return result;
     }
 
-    public void delete(Blog item) {
+    public void delete(Blog item) throws CustomException {
+        if (item == null)
+            throw new CustomException("Parametre null");
+
         if (item != null) {
             deleteById(item.getId());
         }
     }
 
-    public void deleteById(int id) {
+    public void deleteById(Integer id) throws CustomException {
+        if (id == null)
+            throw new CustomException("Parametre null");
 
     }
 }
